Dynamically construct docs example URLs
The links to example tarballs, vagrant, etc all follow the same pattern.
Rather than having to update the docs source on every release, add an
extension that can generate the URLs for these resources dynamically
based on configuration.

Also add a docs environment to tox and fix a couple of warnings with the
docs themselves.