Allow SDK connection if CLAD and major.minor version matches
On CLAD match but version mismatch, check if the major.minor still match (i.e. this is just a patch version change) and if so allow it, but warn user that they should upgrade app or SDK as appropriate.
Added link to the downloads and suggestion that they should download the latest examples when upgrading the SDK.