COZMO-7331 Ensure sync connect calls use same loop
Creating a new loop for every call to _connect_sync breaks the default
connector instance, which maintains a connection to usbmuxd using the
loop first available to it.  This prevents subsequent calls to connect
from succeeding for synchronous functions.

Instead allocate a loop for synchronous code up front and re-use that
for any subsequent sync calls.