COZMO-7567 Get Available Objects before Robot Ready (#42)
Request the available objects earlier, before requesting the available animations - this ensures we have received the available objects before we mark the robot as ready, because we already wait on the animation list to arrive. (Note: adding an additional check to wait for available objects wouldn't be possible without an app change, because there's no guarantee of any message in the case that all cubes and charger are disconnected).