COZMO-6362 Parallel Action Support
Adding in_parallel parameter for all actions, this tells engine to run in parallel (and not clear current actions), and also prevents the SDK from raising an exception if there were pending actions in that case.
OLED face image is now an sent as a parallel action, with previous face actions automatically aborted each time
Aborted actions are immediately not considered in-progress (no longer have to wait a frame for engine to return the cancelled message before starting another action)
Added 06_actions/01_parallel_actions.py example to demonstrate the above.