COZMO-9810 Turn off cube lights on SDK connection
A bug was introduced on the app-side where the EnableLightStates(enable=False) message (which is automatically sent to disable the cube lights on entering SDK mode and at the end of every SDK program) no longer turns off the lights as part of changing the states. This will be resolved for the 1.3 app release, but in the meantime we’ll automatically turn the lights off on cubes when SDK first connects to them each time to avoid cube lights persisting across multiple SDK program runs.