COZMO-8803 Replace name_face Action with Behavior (#104)
Face.name_face was removed in a recent update as the Action to perform this internally was removed. This has been replaced with a Behavior, so the method has been revived to use that.
To enable the use of a Behavior in a way closer to an Action, the behavior code has been updated to track changes in behavior to allow to wait for a behavior to complete, and also catch when a behavior never starts (e.g. if it is not considered runnable if various engine preconditions are not met).
name_face and rename_face now verify that the name string passed in is a valid non-empty ASCII alphabetic-only string (as that’s what the app limits it to).