Switch the box used for Vagrant as old one no longer hosted
boxcutter no longer host boxes, and nobody seems to host a desktop box anymore so switched to an available server box.
Removed vagrant from documentation tabs, and removed the links from the download - the underlying files and the vagrant.html page still exist, so if anyone really needs it or was relying on it they can find it, but it's no longer user-facing so new users won't be confused by it being there (especially now that it's a non-desktop Linux image, so is less user-friendly).